LayoutTests/imported/w3c/web-platform-tests/selection/modify.tentative-expected.txt

Test, these are strong italic normal.
foo
bar
foo
bar
foo


bar

FAIL Stop at previous word boundary when whitespaces are trimmed assert_equals: expected Text node " strong " but got Text node "Test, these are "
PASS Jump linefeed forward
PASS Jump linefeed backward
PASS Jump <br> forward
PASS Jump <br> backward
PASS Jump <br> forward which follows a linefeed
FAIL Jump <br> backward which follows a linefeed assert_equals: expected Text node "foo
" but got Element node <pre id="preLinefeedBr">foo
<br>
bar
</pre>
